requirements.txt
requests
bs4
lxml
behave
PyHamcrest